python/requirements.txt
python-jobspy
pandas
requests
beautifulsoup4
selenium
scrapy
lxml
